Create initial MGC script and add data files
This script compiles into an exact byte-for-byte copy of 20XXTE version 2d rev4.
Compilation requires
[Melee GCI Compiler](https://github.com/dansalvato/melee-gci-compiler/) 1.0.2 or
newer.

However, most of the source code is not complete and only exists in binary form.
Over time, more binary will be replaced with proper source files. The master
branch will always contain a script that compiles exactly into the latest
version of 20XXTE; any merge into master that isn't a version update will be an
improvement of the source files.